requirements.txt
numpy
pygame
rplidar-roboticia
mido
python-rtmidi
